Create variable action aborts if name exists
