+++ src/utils/collectPaths.js
@@ -1,0 +1,68 @@
+import { get, reduce, isArray } from 'lodash';
+
+/**
+ * Collect nodes that match path from `obj`
+ *
+ * Example usage:
+ *
+ * ```
+ * const myObject = {
+ *   stages: [
+ *     {
+ *       prompts: [
+ *         { variable: 'foo' },
+ *         { variable: 'bar' },
+ *       ]
+ *     },
+ *     {
+ *       prompts: [
+ *         { variable: 'bazz' },
+ *       ]
+ *     },
+ *   ]
+ * };
+ *
+ * const paths = collectPaths('stages[].prompts[].variable', myObject);
+ * // result:
+ * // paths = {
+ * //   'stages[0].prompts[0].variable': 'foo',
+ * //   'stages[0].prompts[1].variable': 'bar',
+ * //   'stages[1].prompts[0].variable': 'bazz',
+ * // };
+ * ```
+ *
+ * @returns {object} in format: { [path]: node }
+ */
+const collectPaths = (paths, obj, memoPath) => {
+  // We expect a string notation for paths, but it's actually converted to array automatically:
+  // 'stages[].prompts[].subject.type' => ['stages', 'prompts', 'subject.type']
+  const [next, ...rest] = isArray(paths) ?
+    paths :
+    paths.split('[].');
+
+  const path = memoPath ?
+    `${memoPath}.${next}` : `${next}`;
+
+  const nextObj = get(obj, next);
+
+  if (rest.length > 0) {
+    return reduce(
+      nextObj || [],
+      (memo, item, index) => ({
+        ...memo,
+        ...collectPaths(rest, item, `${path}[${index}]`),
+      }),
+      {},
+    );
+  }
+
+  if (nextObj) {
+    return {
+      [path]: nextObj,
+    };
+  }
+
+  return {};
+};
+
+export default collectPaths;
